+++ demos/bevy/client/src/app.rs
@@ -14,7 +14,7 @@ use bevy_winit::WinitPlugin;
 use naia_bevy_client::{ClientConfig, Plugin as ClientPlugin};
 use naia_bevy_demo_shared::protocol;
 
-use crate::systems::{events, init, input, sync, tick_events};
+use crate::systems::{events, init, input, sync};
 
 pub fn run() {
     App::default()
@@ -43,7 +43,7 @@ pub fn run() {
         .add_system_to_stage(CoreStage::PreUpdate, events::insert_component_events)
         .add_system_to_stage(CoreStage::PreUpdate, events::update_component_events)
         .add_system_to_stage(CoreStage::PreUpdate, events::message_events)
-        .add_system_to_stage(CoreStage::PreUpdate, tick_events)
+        .add_system_to_stage(CoreStage::PreUpdate, events::tick_events)
         .add_system_to_stage(CoreStage::Update, input)
         .add_system_to_stage(CoreStage::Update, sync)
         // Run App
