+++ demos/bevy/client/src/systems/mod.rs
@@ -3,9 +3,7 @@ pub mod events;
 mod init;
 mod input;
 mod sync;
-mod tick;
 
 pub use init::init;
 pub use input::input;
 pub use sync::sync;
-pub use tick::tick_events;
