+++ demos/bevy/server/src/resources.rs
@@ -4,8 +4,6 @@ use bevy_ecs::{entity::Entity, prelude::Resource};
 
 use naia_bevy_server::{RoomKey, UserKey};
 
-use naia_bevy_demo_shared::messages::KeyCommand;
-
 #[derive(Resource)]
 pub struct Global {
     pub main_room_key: RoomKey,
