+++ demos/bevy/server/src/systems/init.rs
@@ -33,6 +33,5 @@ pub fn init(mut commands: Commands, mut server: Server) {
     commands.insert_resource(Global {
         main_room_key,
         user_to_entity_map: HashMap::new(),
-        player_last_command: HashMap::new(),
     })
 }
