+++ demos/bevy/server/src/systems/mod.rs
@@ -1,7 +1,5 @@
 pub mod events;
 
 mod init;
-mod tick;
 
 pub use init::init;
-pub use tick::tick_events;
